services/document_service/requirements.txt
# Copyright (c) 2025 LALO AI SYSTEMS, LLC. All rights reserved.
#
# PROPRIETARY AND CONFIDENTIAL
#
# This file is part of LALO AI Platform and is protected by copyright law.
# Unauthorized copying, modification, distribution, or use of this software,
# via any medium, is strictly prohibited without the express written permission
# of LALO AI SYSTEMS, LLC.
#

fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==1.4.23
aiofiles==0.7.0
python-docx==0.8.11
openpyxl==3.0.9
pypdf
azure-storage-blob==12.9.0
msal==1.16.0
office365-rest-python-client==2.3.1
